requirements.txt
python-telegram-bot==20.4
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
